+++ HSTracker/Logging/CoreManager.swift
@@ -669,17 +669,17 @@ final class CoreManager: NSObject {
             playerClass = DefaultDecks.DungeonRun.getUldumHeroPlayerClass(playerClass: loadout.playerClass)
         } else if isPVPDR {
             if info.heroClass.intValue != 0 {
-                playerClass = CardClass.allCases[info.heroClass.intValue]
+                playerClass = CardClass.allCases[safeIndex: info.heroClass.intValue] ?? .invalid
 
             } else if info.heroCardDbId.intValue != 0, let cc = tryGetHeroClass(dbfId: info.heroCardDbId.intValue) {
                 playerClass = cc
             } else if info.playerSelectedHeroDbId.intValue != 0, let cc = tryGetHeroClass(dbfId: info.playerSelectedHeroDbId.intValue) {
                 playerClass = cc
             } else if info.heroCardClass.intValue != 0 {
-                playerClass = CardClass.allCases[info.heroCardClass.intValue]
+                playerClass = CardClass.allCases[safeIndex: info.heroCardClass.intValue] ?? .invalid
             }
         } else {
-            playerClass = CardClass.allCases[info.heroClass.intValue != 0 ? info.heroClass.intValue : info.heroCardClass.intValue]
+            playerClass = CardClass.allCases[safeIndex: info.heroClass.intValue != 0 ? info.heroClass.intValue : info.heroCardClass.intValue] ?? .invalid
         }
         var deck = RealmHelper.getDecks()?.filter({ x in x.isActive && (!isPVPDR && x.isDungeon || isPVPDR && x.isDuels)
                                                     &&  x.playerClass == playerClass
@@ -817,7 +817,7 @@ final class CoreManager: NSObject {
         
         let ret = Deck()
         ret.name = deck.title
-        ret.heroId = CardClass.allCases[deck.clazz.intValue].defaultHeroCardId
+        ret.heroId = (CardClass.allCases[safeIndex: deck.clazz.intValue] ?? .invalid).defaultHeroCardId
         
         let tmpCards = Dictionary(grouping: deck.cards, by: { x in x }).compactMap { (key: NSNumber, value: [NSNumber]) -> RealmCard? in
             guard let card = Cards.by(dbfId: key.intValue, collectible: false) else {