+++ HSTracker/UIs/Battlegrounds/Guides/Heroes/BattlegroundsHeroGuideViewModel.swift
@@ -35,8 +35,7 @@ struct BattlegroundsHeroGuideViewModel {
 
         let availableRaces = Set(AppDelegate.instance().coreManager.game.availableRaces ?? [])
         self.favorableTribes = (heroGuide?.favorable_tribes ?? []).compactMap { raceNumber -> Race? in
-            guard raceNumber >= 0, raceNumber < Race.allCases.count else { return nil }
-            let race = Race.allCases[raceNumber]
+            guard let race = Race.allCases[safeIndex: raceNumber] else { return nil }
             return availableRaces.contains(race) ? race : nil
         }
     }